+++ q2_types/plugin_setup.py
@@ -35,4 +35,4 @@ importlib.import_module('q2_types.ordination')
 importlib.import_module('q2_types._sample_data')
 importlib.import_module('q2_types.feature_data')
 importlib.import_module('q2_types.reference_features')
-importlib.import_module('q2_types._per_sample_sequences')
+importlib.import_module('q2_types.per_sample_sequences')
